pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "notifyme-bot"
version = "0.1.0"
description = "Telegram reminder bot with SQLite and pluggable LLM backends."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-telegram-bot>=22.5,<23.0",
  "apscheduler>=3.10.4,<4.0",
  "aiosqlite>=0.20.0",
  "httpx>=0.27.0",
  "PyYAML>=6.0.2",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.11.0",
  "pytest>=8.3.5",
  "pytest-asyncio>=0.25.3",
  "pytest-cov>=6.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
notifyme_bot = ["locales/*.yaml"]

[tool.ruff]
line-length = 79
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N"]

[tool.ruff.lint.per-file-ignores]
"src/notifyme_bot/llm_parser.py" = ["E501"]

[tool.pytest.ini_options]
addopts = [
  "--cov=notifyme_bot",
  "--cov-report=term-missing",
  "--cov-fail-under=85",
]
pythonpath = ["tests"]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
branch = true
source = ["notifyme_bot"]
omit = ["*/__main__.py"]

[tool.coverage.report]
show_missing = true
precision = 2
